create monk shared constants
