+++ src/analysis/retail/monk/mistweaver/constants.ts
@@ -4,50 +4,6 @@ import { TALENTS_MONK } from 'common/TALENTS';
 import { Talent } from 'common/TALENTS/types';
 import Combatant from 'parser/core/Combatant';
 
-export const ABILITIES_AFFECTED_BY_HEALING_INCREASES = [
-  // Spells
-  TALENTS_MONK.ENVELOPING_MIST_TALENT.id,
-  SPELLS.ENVELOPING_MIST_TFT.id,
-  SPELLS.RENEWING_MIST_CAST.id,
-  SPELLS.RENEWING_MIST_HEAL.id,
-  SPELLS.VIVIFY.id,
-  SPELLS.INVIGORATING_MISTS_HEAL.id,
-  TALENTS_MONK.SOOTHING_MIST_TALENT.id,
-  SPELLS.SOOTHING_BREATH.id,
-  SPELLS.AT_CRIT_HEAL.id,
-  SPELLS.AT_HEAL.id,
-  SPELLS.GUSTS_OF_MISTS.id,
-  SPELLS.GUST_OF_MISTS_CHIJI.id,
-  SPELLS.EXPEL_HARM.id,
-  SPELLS.WOTC_HEAL.id,
-  SPELLS.WOTC_CRIT_HEAL.id,
-
-  //Hero Talents
-  TALENTS_MONK.CELESTIAL_CONDUIT_MISTWEAVER_TALENT.id,
-  SPELLS.FLIGHT_OF_THE_RED_CRANE_HEAL.id,
-  SPELLS.FLIGHT_OF_THE_RED_CRANE_UNITY.id,
-  SPELLS.MANTRA_OF_PURITY_HEAL.id,
-  SPELLS.COURAGE_OF_THE_WHITE_TIGER_HEAL.id,
-  SPELLS.HARMONIC_SURGE_HEAL.id,
-
-  // Cooldowns
-  TALENTS_MONK.REVIVAL_TALENT.id,
-  TALENTS_MONK.RESTORAL_TALENT.id,
-
-  // Talents
-  TALENTS_MONK.CHI_WAVE_TALENT.id,
-  TALENTS_MONK.SHEILUNS_GIFT_TALENT.id,
-  SPELLS.RISING_MIST_HEAL.id,
-  TALENTS_MONK.HEALING_WINDS_TALENT.id,
-  SPELLS.HEALING_WINDS_HEAL.id,
-  SPELLS.ZEN_PULSE_HEAL.id,
-  SPELLS.YULONS_WHISPER_HEAL.id,
-  SPELLS.BURST_OF_LIFE_HEAL.id,
-  SPELLS.OVERFLOWING_MISTS_HEAL.id,
-  // Misc
-  SPELLS.LEECH.id,
-];
-
 export const THUNDER_FOCUS_TEA_SPELLS = [
   TALENTS_MONK.RISING_SUN_KICK_TALENT,
   TALENTS_MONK.RUSHING_WIND_KICK_MISTWEAVER_TALENT,
@@ -64,7 +20,6 @@ export const AT_TRANSFER_COEFFICIENT = 0.25;
 export const AT_MAX_TARGETS = 5;
 
 // Talent Constants
-export const SAVE_THEM_ALL_MAX_INCREASE = 0.1;
 export const AMPLIFIED_RUSH_HEALING_BOOST = 0.5;
 export const JADE_EMPOWERMENT_DAMAGE_BOOST = 3.0;
 export const JADE_EMPOWERMENT_SECONDARY_TARGET_EFFECTIVENESS = 0.25;
